修复openLoopSteering_Handler函数参数类型，优化PID控制器的积分计算，调整任务3的旋转角度和速度设置

+++ control/CONTROL/control.h
@@ -99,7 +99,7 @@ void SteeringTest_CyclicRotation(void);
 void Test_Handler(void);
 void normal_Handler(void); 
 int moveForward_Handler(void);
-u8 openLoopSteering_Handler(int SteerTime,u8 PWM_Value);
+u8 openLoopSteering_Handler(int SteerTime,int PWM_Value);
 
 // 转向控制参数调节函数
 void Set_Steering_Kp(u16 kp);
